% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_to_crit.R
\name{gamma_to_crit}
\alias{gamma_to_crit}
\title{Transforms gamma posterior samples to midpoint centered criteria.}
\usage{
gamma_to_crit(samples, beta_index = 1, gamma_link = "softmax", s = 2)
}
\arguments{
\item{samples}{a data frame of posterior samples from stanfit object (e.g.,
samples = as.data.frame(stanfit)).}

\item{beta_index}{a fixed effect index (a scalar) = an index of a
\code{gamma_fixed} matrix column. See also 'Details' for which columns of
the \code{gamma_fixed} matrix can and which cannot be converted to criteria
using this function.}

\item{gamma_link}{is either 'softmax' (described in the paper), 'log_distance' or 'log_ratio'
(See the Readme file in the github repository)}

\item{s}{a criteria scaling factor (a scalar). Warning: must be equal to the
value used when fitting the model.}
}
\value{
a data frame containing posterior samples of midpoint centered
  criteria.
}
\description{
\code{gamma_to_crit} transforms gamma posterior samples to midpoint centered
criteria. This will work correctly only if the column of the gamma fixed
effects parameter matrix specified by \code{beta_index} represents values of
gamma for a given condition, not the regression slope or the difference in
gamma values between conditions (see also 'Details').
}
\details{
In general, in an SDT model there is only one sensitivity but there can be
more than one criterion. When there is more than one criterion the regression
coefficients that represent the dependence of criteria on additional
variables form a matrix, not a vector. Row j of this fixed effects matrix
contains regression coefficients for the j-th criterion. Let's say we have
data from a simple experiment with experimental and control groups and we
want to test if criteria depend on group membership. If we specify the fixed
effects model matrix by calling \code{make_stan_data} with \code{fixed =
list(gamma = ~ condition, ...)} then the first column of the
\code{gamma_fixed} parameter matrix will contain the intercepts for every
criterion (in the gamma space), which is the value of gamma in the condition
chosen as base level, and the second column will contain the difference in
gamma values between experimental and control groups. In this case
\code{gamma_to_crit(samples, 1)} will work, but \code{gamma_to_crit(samples,
2)} will not work. If, on the other hand, we wanted to work with the actual
criteria, not the gamma values from which criteria are derived, then
\code{fixed = list(gamma = ~ -1 + condition, ...)} would allow us to apply
the \code{gamma_to_crit} function to both columns of the \code{gamma_fixed}
matrix because \code{~ -1 + condition} results in a separate intercepts and
slopes parametrization of the gamma fixed effects model matrix. This way we
could easily estimate the criteria for both conditions.
}
\examples{
## First we apply the inverse of what happens inside the gamma_to_crit function
criteria = c(-2, -1, 0, 1, 2)
cumprobs = pnorm(criteria)
cumprobs = c(pnorm(criteria, sd = 2), 1)
areas = c(cumprobs[1], cumprobs[-1] - cumprobs[-length(cumprobs)])
gamma = log(areas / areas[length(areas)])
## than we apply the simplified version of gamma_to_crit
g_to_c = function(x, s = 2){
  x = exp(x)
  s * qnorm(cumsum(x/sum(x))[-length(x)])
}
## so that we can see that all is well...
rbind(g_to_c(gamma), criteria)
}
