% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_to_thresholds.R
\name{gamma_to_thresholds}
\alias{gamma_to_thresholds}
\title{Transforms gamma posterior samples or contrasts of gamma posterior samples to thresholds}
\usage{
gamma_to_thresholds(fit, gamma_link, ..., s = 2, K = NULL)
}
\description{
Transforms gamma posterior samples or contrasts of gamma posterior samples to thresholds
}
