% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data.R
\name{make_stan_data}
\alias{make_stan_data}
\title{Creates the list of data structures required by the stan function.}
\usage{
make_stan_data(
  adata,
  fixed,
  random = list(),
  criteria_scale = 2,
  metad = FALSE
)
}
\arguments{
\item{adata}{an aggregated data object created by
\code{aggregate_responses} function.}

\item{fixed}{a list specifying gamma and delta fixed effects and
priors. See also 'Details'.}

\item{random}{an optional list specifying gamma and delta random
effects and priors. The default is list(), which corresponds to
a non-hierarchical SDT model. See also 'Details'.}

\item{criteria_scale}{a scaling factor corresponding to mapping
distribution's standard deviation. The default is 2. See also
'Details'.}

\item{metad}{when TRUE additional data required by the meta-d'
model are created, the default if FALSE.}
}
\value{
a list with response and stimulus data, model matrices,
    prior parameter values, and other data required by the stan
    model generated using \code{make_stan_model}.
}
\description{
\code{make_stan_data} creates the list of data structures required by the
\code{stan} function when fitting the model generated by
\code{make_stan_model}.
}
\details{
\code{fixed} must be a list of model formulae. It may also contain prior
parameter values, if non-default priors on fixed effects are required. This
list is composed of the following elements:

\describe{

\item{delta}{is a model formula that defines the delta fixed effects model
matrix, e.g., \code{delta = ~ condition}}

\item{gamma}{is a model formula that defines the gamma fixed effects model
matrix.}

\item{delta_mu}{is na optional vector specifying means of independent normal
priors on delta fixed effects. This must be of the same length as the number
of delta fixed effects or of length 1, in which case the same value will be
used for every delta fixed effect. The default value is
\code{acc_to_delta(.75)}}

\item{delta_sd}{is na optional vector specifying standard deviations of
independent normal priors on delta fixed effects. This must be of the same
length as the number of delta fixed effects or of length 1, in which case the
same value will be used for every delta fixed effect. The default value is
\code{.5 * (acc_to_delta(.99) - acc_to_delta(.51))}}

\item{gamma_mu}{is na optional vector specifying means of independent normal
priors on gamma fixed effects. This must be of the same length as the number
of gamma fixed effects or of length 1, in which case the same value will be
used for every gamma fixed effect. Note that when there is more than one
criterion the number of gamma fixed effects is equal to the number I of
columns of gamma fixed effects model matrix times the number of criteria C.
Internally, \code{gamma_mu} is stored as a CxI matrix (in column major
order): element C*j+i of the \code{gamma_mu} vector (where 0 < j and 0 < i <
C) represents fixed effect corresponding to the i-th column of the gamma
fixed effects model matrix and the j-th criterion. The default value is 0.}

\item{gamma_sd}{is na optional vector specifying standard deviations of
independent normal priors on gamma fixed effects. This must be of the same
length as the number of gamma fixed effects or of length 1, in which case the
same value will be used for every gamma fixed effect. See \code{gamma_mu}
above for details on how the elements of the \code{gamma_sd} vector
correspond to elements of the gamma fixed effects parameter matrix. The
default value is log(100) which means that a priori areas under the
gamma-to-criterion mapping distribution curve delineated by the criteria are
expected to vary by a factor of 100 or less but varying by a factor of
\code{exp(2*log(100)) = 10000} is highly unlikely.}

}

\code{random} is on optional list of lists of model formulae and prior
parameter values, if non-default priors on random effects are required. Each
list specifies delta and gamma random effects of one grouping factor:

\describe{

\item{group}{is a model formula specifying the random grouping factor, e.g.,
\code{group = ~ subject}}

\item{delta}{is a model formula that defines the delta random effects model
matrix. This must define a submodel of the delta fixed effects model, e.g.,
if delta depends on condition and condition is a within-subject variable then
\code{delta = ~ condition} is a valid delta random effects specification.}

\item{gamma}{is a model formula that defines the gamma random effects model
matrix. This must define a submodel of the gamma fixed effects model, e.g.,
if gamma depends on condition and condition is a within-subject variable then
\code{gamma = ~ condition} is a valid gamma random effects specification.}

\item{delta_nu}{is an optional vector of lkj prior parameters for delta
random effects. This must be of the same length as the number of delta random
effects or of length 1, in which case the same value will be used for every
delta random effect. The default is 1 which corresponds to uniform prior on
random effects correlation matrices. The greater the value of this parameter
the more emphasis is put on zer off-diagonal correlations.}

\item{delta_sd_scale}{is an optional vector of half-Cauchy prior parameters
for delta random effects. This must be of the same length as the number of
delta random effects or of length 1, in which case the same value will be
used for every delta random effect. The default value is \code{.5 *
(acc_to_delta(.99) - acc_to_delta(.51))}}

\item{gamma_nu}{is an optional vector of lkj prior parameters for gamma
random effects. This must be of the same length as the number of gamma random
effects or of length 1, in which case the same value will be used for every
gamma random effect. The default is 1 which corresponds to uniform prior on
random effects correlation matrices. The greater the value of this parameter
the more emphasis is put on zer off-diagonal correlations. See
\code{gamma_mu} above for details on how the elements of the \code{gamma}
vector correspond to elements of the gamma parameter matrix.}

\item{gamma_sd_scale}{is an optional vector of half-Cauchy prior parameters
for gamma random effects. This must be of the same length as the number of
gamma random effects or of length 1, in which case the same value will be
used for every gamma random effect. The default value is \code{log(100)}. See
\code{gamma_mu} above for details on how the elements of the \code{gamma}
vector correspond to elements of the gamma parameter matrix.}

}
}
\examples{
data(gabor)
gabor$r = combined_response(gabor$stim, gabor$rating, gabor$acc)
adata = aggregate_responses(gabor, 'stim', 'r', c('duration', 'id', 'order'))
fixed = list(delta = ~ -1 + duration:order, gamma = ~ order)
random = list(list(group = ~ id, delta = ~ -1 + duration, gamma = ~ 1))
sdata = make_stan_data(adata, fixed, random)
sdata
}
