% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_model.R
\name{make_stan_model}
\alias{make_stan_model}
\title{Creates the SDT stan model code.}
\usage{
make_stan_model(
  random = NULL,
  gamma_link = "softmax",
  model = "sdt",
  delta_link = "log"
)
}
\arguments{
\item{random}{an optional list specifying random effects
structure. See \code{\link{make_stan_data}} for details.}

\item{gamma_link}{can be either 'softmax' (described in the paper),
'log_distance', or 'log_ratio' (See the Readme file in the
github repository)}

\item{model}{can be either 'sdt' (the default), 'uvsdt', 'metad',
'ordinal', or 'uvordinal'}
}
\value{
a string containing the full model definition in the stan
    modelling language.
}
\description{
\code{make_stan_model} function creates the stan model code
defining the SDT model with additional regression structure and
random effects if these are specified using the \code{random}
argument. See \code{\link{make_stan_data}} for details on how to
specify a random effects structure. The model is described in detail
in the paper (TODO reference)
}
\examples{
data(gabor)
model = make_stan_model(list(list(group = ~ id, delta = ~ -1 + duration, gamma = ~ 1)))
cat(model)
}
