% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdt_fit.R
\name{plot_sdt_fit}
\alias{plot_sdt_fit}
\title{Creates ROC or response distribution model fit plots with posterior
predictive intervals.}
\usage{
plot_sdt_fit(
  fit,
  adata,
  variables = NULL,
  type = "roc",
  alpha = 0.05,
  bw = TRUE
)
}
\arguments{
\item{fit}{a stanfit object.}

\item{adata}{an aggregated data object produced by the
\code{aggregate_responses} function. It must be the same
aggregated data object that was used to produce the stanfit
object.}

\item{variables}{an optional vector of names of variables. These
variables, if specified, define the subsets of data that will
be represented by separate plots.}

\item{type}{a type of plot required. If type = 'roc' then ROC curve
plots are created, if type != 'roc' then response distribution
plots are created.}

\item{alpha}{an alpha level for the posterior predictive intervals,
e.g., if alpha = .05 then 100\% - alpha = 95\% posterior
predictive intervals will be calculated.}

\item{bw}{if TRUE (the default), a black-and-white publication-friendly version
of the response distribution plot will be created.}
}
\value{
a plot object.
}
\description{
Creates ROC or response distribution model fit plots with posterior
predictive intervals.
}
